테마 모드 스위치 부분에 대한 리팩토링 추가
