--- src/globals/DefaultTheme.tsx
@@ -1,26 +1,0 @@
-type DefaultTheme = {
-  basicBg: string;
-  basicBtnBg: string;
-  secondBg: string;
-  basicFont: string;
-  basicContentBox: string;
-  basicLine: string;
-};
-
-export const darkTheme: DefaultTheme = {
-  basicBg: "#1E2022",
-  basicBtnBg: "#3C5186",
-  secondBg: "#6F9EAF",
-  basicFont: "#ddd",
-  basicContentBox: "#828282",
-  basicLine: "#ccc",
-};
-
-export const lightTheme: DefaultTheme = {
-  basicBg: "#fff",
-  basicBtnBg: "#E1F2FB",
-  secondBg: "#ecdbba",
-  basicFont: "#000",
-  basicContentBox: "#fff",
-  basicLine: "#333",
-};
