+++ src/globals/theme.ts
@@ -1,0 +1,19 @@
+export type DefaultTheme = typeof lightTheme;
+
+export const lightTheme = {
+  basicBg: "#fff",
+  basicBtnBg: "#E1F2FB",
+  secondBg: "#ecdbba",
+  basicFont: "#000",
+  basicContentBox: "#fff",
+  basicLine: "#333",
+};
+
+export const darkTheme: DefaultTheme = {
+  basicBg: "#1E2022",
+  basicBtnBg: "#3C5186",
+  secondBg: "#6F9EAF",
+  basicFont: "#ddd",
+  basicContentBox: "#828282",
+  basicLine: "#ccc",
+};
